Finish blueprint migration and test routes
