+++ requirements.txt
@@ -11,7 +11,6 @@ Flask-Limiter==3.10.1
 Flask-SQLAlchemy==3.1.1
 Flask-WTF==1.2.2
 greenlet==3.1.1
-gunicorn==23.0.0
 itsdangerous==2.2.0
 Jinja2==3.1.4
 limits==4.0.1
@@ -25,10 +24,10 @@ pycparser==2.22
 Pygments==2.19.1
 PyJWT==2.10.0
 PyMySQL==1.1.1
+python-dotenv==1.1.0
 rich==13.9.4
 SQLAlchemy==2.0.36
 typing_extensions==4.12.2
 Werkzeug==3.1.3
 wrapt==1.17.2
 WTForms==3.2.1
-python-dotenv==1.1.0
