+++ server/.env
@@ -1,1 +1,8 @@
 FLASK_SECRET_KEY=3f368e4bf7f7e2ef7fe58d6244cb884a
+JWT_SECRET_KEY=secret_key
+DB_USER=server_user
+DB_PASSWORD=server_password
+DB_NAME=banking_db_v0
+DB_HOST=localhost
+SECRET_KEY=supersecretkey123
+
